requirements/base.txt
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
dj-pagination==2.4.0
Django==2.2.5
django-admin-thumbnails==0.2.5
django-appconf==1.0.4
django-ckeditor==5.9.0
django-environ==0.4.5
django-filter==2.2.0
django-imagekit==4.0.2
django-import-export==2.0.2
django-jet==1.0.8
django-js-asset==1.2.2
django-modeltranslation==0.14.4
django-rest-swagger==2.2.0
djangorestframework==3.11.0
drf-yasg==1.17.1
et-xmlfile==1.0.1
fcm-django==0.3.4
future==0.18.2
huey==2.2.0
idna==2.9
inflection==0.4.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
Markdown==3.2.1
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.3
packaging==20.3
pilkit==2.0
Pillow==7.1.1
psycopg2-binary==2.8.4
pyfcm==1.4.7
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.2
PyYAML==5.3.1
redis==3.5.0
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.12.0
sqlparse==0.3.1
tablib==1.1.0
uritemplate==3.0.1
urllib3==1.25.8
xlrd==1.2.0
xlwt==1.3.0
